Commit do desespero

+++ src/kine_controller.cpp
@@ -1,5 +1,5 @@
 #include <ros/ros.h>
-#include <std_msgs/Float64.h>
+#include <std_msgs/Float32.h>
 #include <geometry_msgs/Twist.h>
 
 ros::Publisher pub1;
@@ -18,13 +18,13 @@ int main(int argc, char **argv)
     double r= 0.1,b = 0.35;
     ros::init(argc,argv,"ros_Ctrl");
     ros::NodeHandle node;
-    pub1 = node.advertise<std_msgs::Float64>("vrep/vehicle/motorLeftSpeed",1);
-    pub2 = node.advertise<std_msgs::Float64>("vrep/vehicle/motorRightSpeed",1);
+    pub1 = node.advertise<std_msgs::Float32>("vrep/vehicle/motorLeftSpeed",1);
+    pub2 = node.advertise<std_msgs::Float32>("vrep/vehicle/motorRightSpeed",1);
     ros::Rate loop_rate(60);
     
     ros::Subscriber sub_avd = node.subscribe<geometry_msgs::Twist>("ros_odom",1,callback);
-    std_msgs::Float64 w_right;
-    std_msgs::Float64 w_left;
+    std_msgs::Float32 w_right;
+    std_msgs::Float32 w_left;
     
     while( ros::ok())
     {
